code/requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
CacheControl==0.12.6
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.2
daphne==2.5.0
distlib==0.3.0
distro==1.4.0
Django==3.0.7
django-cors-headers==3.2.1
django-environ==0.4.5
django-model-utils==4.0.0
djangorestframework==3.11.0
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
mysqlclient==1.4.6
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
pkg-resources==0.0.0
progress==1.5
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-decouple==3.3
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
webencodings==0.5.1
zope.interface==5.1.0
